Fix removal of wanted files
Previous implementation of `400-clean` build script was expecting a very specific directory structure of `repo/addons`, or otherwise it could be removing things that you actually wanted.

Imagine for instance that 2 doodba projects share the same private addons, so you add them in `odoo/src/private` in the 1st one, and link them in the 2nd one with repos+addons combination, adding something like this to `addons.yaml`:

```yaml
other-doodba/odoo/src/private:
- custom_other_addon
```

Previous implementation of the clean step would be removing the `other-doodba` folder entierly, as it was not following the expected structure.

This new implementation is smarter:

1. Preserve all Odoo source code, outside of its `addons` folder (which is handled as a normal addons repo, to be able to restrict them if wanted).
2. For other folders, check if it can be found inside an activated addons path. If not, remove it.
3. Repeat, recursively, bottom-up.

It should be more flexible and less error-prone. A test that was testing the cleanup has been slightly modified to support this use case.